finish routing solver output

+++ open-trip-planner/src/main/scala/com/azavea/busplan/routing/RouteVertex.scala
@@ -1,4 +1,4 @@
 package com.azavea.busplan.routing
 
 case class RouteVertex(route: String, routeSequence: Int, stopSequence: Int,
-  time: Int, x: Double, y: Double)
+  time: Long, x: Double, y: Double)
